// core/src/data-table/main/foundation/table.api.ts
export type RowKeyGetter<RowType> = (params: { row: RowType; rowIndex?: number }) => Key;